refactor: move up events to own file under analytics pkg
